requirements.txt
# Core dependencies for RAG AI Agent
pydantic>=2.0.0
pydantic-ai>=0.0.13
openai>=1.0.0
supabase>=2.0.0
streamlit>=1.30.0
python-dotenv>=1.0.0

# Document processing
PyPDF2>=3.0.0
python-docx>=1.1.0

# Vector operations and similarity
numpy>=1.24.0
scikit-learn>=1.3.0

# Database and embeddings
psycopg2-binary>=2.9.0
pgvector>=0.2.0

# Visualization and analytics
plotly>=5.0.0
pandas>=2.0.0

# HTTP client
httpx>=0.25.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development tools
black>=23.0.0
isort>=5.12.0

# Logging and observability
loguru>=0.7.0

# Additional utilities
tqdm>=4.65.0
click>=8.1.0